Split SimpleKafkaSpecExecutorInstance into Producer and Consumer
